backend/requirements.txt
# IBMI Backend - dependências principais
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0
pydantic-settings>=2.0
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30.0
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.9
email-validator>=2.0
# Necessário para o Alembic ler o .env fora do contexto FastAPI
python-dotenv>=1.0.0
